--- debian/webbrowser-app-apparmor.manifest
@@ -1,43 +1,0 @@
-{
-  "profiles": {
-    "webbrowser-app": {
-      "binary": "/usr/bin/webbrowser-app",
-      "profile_name": "webbrowser-app",
-      "policy_vendor": "ubuntu",
-      "policy_version": 1.3,
-      "policy_groups": [
-        "accounts",
-        "audio",
-        "camera",
-        "content_exchange",
-        "content_exchange_source",
-        "keep-display-on",
-        "location",
-        "microphone",
-        "networking",
-        "push-notification-client",
-        "video",
-        "webview"
-      ],
-      "abstractions": [
-        "ibus",
-        "nvidia",
-        "user-tmp"
-      ],
-      "template_variables": {
-        "APP_ID_DBUS": "webbrowser_2dapp",
-        "APP_PKGNAME_DBUS": "webbrowser_2dapp",
-        "APP_PKGNAME": "webbrowser-app",
-        "CLICK_DIR": "/usr/share/webbrowser-app"
-      },
-      "read_path": [
-        "/usr/share/applications/",
-        "/custom/vendor/here/location-provider/consent/*.html",
-        "@{HOME}/.local/share/applications/",
-        "/sys/devices/**/input*/"
-      ],
-      "write_path": [
-      ]
-    }
-  }
-}
