--- debian/webbrowser-app-autopilot.install
@@ -1,1 +1,0 @@
-usr/lib/python3/dist-packages/webbrowser_app
