--- debian/webbrowser-app.install
@@ -1,11 +1,0 @@
-usr/bin/webbrowser-app
-usr/share/webbrowser-app/*.qml
-usr/share/webbrowser-app/*.js
-usr/share/webbrowser-app/*.png
-usr/share/webbrowser-app/actions/*.qml
-usr/share/webbrowser-app/webbrowser/
-usr/share/applications/webbrowser-app.desktop
-usr/share/locale/*/LC_MESSAGES/webbrowser-app.mo
-usr/share/url-dispatcher/
-usr/share/content-hub/peers/webbrowser-app
-debian/usr.bin.webbrowser-app etc/apparmor.d
