--- debian/webbrowser-app.lintian-overrides
@@ -1,1 +1,0 @@
-webbrowser-app: binary-without-manpage usr/bin/webbrowser-app
