+++ src/app/CMakeLists.txt
@@ -40,10 +40,10 @@ target_link_libraries(${COMMONLIB}
 )
 
 file(GLOB QML_FILES *.qml)
-install(FILES ${QML_FILES} DESTINATION ${CMAKE_INSTALL_DATADIR}/webbrowser-ng)
+install(FILES ${QML_FILES} DESTINATION ${CMAKE_INSTALL_DATADIR}/morph-browser)
 file(GLOB JS_FILES *.js)
-install(FILES ${JS_FILES} DESTINATION ${CMAKE_INSTALL_DATADIR}/webbrowser-ng)
-install(DIRECTORY actions DESTINATION ${CMAKE_INSTALL_DATADIR}/webbrowser-ng
+install(FILES ${JS_FILES} DESTINATION ${CMAKE_INSTALL_DATADIR}/morph-browser)
+install(DIRECTORY actions DESTINATION ${CMAKE_INSTALL_DATADIR}/morph-browser
         FILES_MATCHING PATTERN *.qml)
 
 add_subdirectory(webbrowser)
