+++ src/app/webbrowser/CMakeLists.txt
@@ -1,4 +1,4 @@
-project(webbrowser-ng)
+project(morph-browser)
 
 find_package(Qt5Concurrent REQUIRED)
 find_package(Qt5Sql REQUIRED)
@@ -23,7 +23,7 @@ set(WEBBROWSER_APP_MODELS_SRC
     text-search-filter-model.cpp
 )
 
-set(WEBBROWSER_APP_MODELS webbrowser-ng-models)
+set(WEBBROWSER_APP_MODELS morph-browser-models)
 
 add_library(${WEBBROWSER_APP_MODELS} STATIC ${WEBBROWSER_APP_MODELS_SRC})
 target_link_libraries(${WEBBROWSER_APP_MODELS}
@@ -37,10 +37,10 @@ set(WEBBROWSER_APP_SRC
     file-operations.cpp
     reparenter.cpp
     searchengine.cpp
-    webbrowser-ng.cpp
+    morph-browser.cpp
 )
 
-set(WEBBROWSER_APP webbrowser-ng)
+set(WEBBROWSER_APP morph-browser)
 
 add_executable(${WEBBROWSER_APP} ${WEBBROWSER_APP_SRC})
 target_link_libraries(${WEBBROWSER_APP}
@@ -56,14 +56,14 @@ install(TARGETS ${WEBBROWSER_APP}
         RUNTIME DESTINATION ${CMAKE_INSTALL_BINDIR})
 
 file(GLOB QML_FILES *.qml qmldir *.js)
-install(FILES ${QML_FILES} DESTINATION ${CMAKE_INSTALL_DATADIR}/webbrowser-ng/webbrowser)
+install(FILES ${QML_FILES} DESTINATION ${CMAKE_INSTALL_DATADIR}/morph-browser/webbrowser)
 
 install(DIRECTORY assets
-        DESTINATION ${CMAKE_INSTALL_DATADIR}/webbrowser-ng/webbrowser
+        DESTINATION ${CMAKE_INSTALL_DATADIR}/morph-browser/webbrowser
         FILES_MATCHING PATTERN *.png PATTERN *.svg PATTERN *.sci)
 
 install(DIRECTORY searchengines
-        DESTINATION ${CMAKE_INSTALL_DATADIR}/webbrowser-ng/webbrowser
+        DESTINATION ${CMAKE_INSTALL_DATADIR}/morph-browser/webbrowser
         FILES_MATCHING PATTERN *.xml)
 
 configure_file(${DESKTOP_FILE}.in.in ${DESKTOP_FILE}.in @ONLY)
@@ -75,12 +75,12 @@ add_custom_target(${DESKTOP_FILE} ALL
 install(FILES ${CMAKE_CURRENT_BINARY_DIR}/${DESKTOP_FILE}
         DESTINATION ${CMAKE_INSTALL_DATADIR}/applications)
 
-install(FILES "webbrowser-ng.url-dispatcher"
+install(FILES "morph-browser.url-dispatcher"
         DESTINATION ${CMAKE_INSTALL_DATADIR}/url-dispatcher/urls)
 
-install(FILES "webbrowser-ng-content-hub.json"
+install(FILES "morph-browser-content-hub.json"
     DESTINATION ${CMAKE_INSTALL_DATADIR}/content-hub/peers
-    RENAME webbrowser-ng
+    RENAME morph-browser
     )
 
 if(${CLICK_MODE})
@@ -94,5 +94,5 @@ if(${CLICK_MODE})
 
   configure_file(manifest.json.in ${CMAKE_CURRENT_BINARY_DIR}/manifest.json)
   install(FILES ${CMAKE_CURRENT_BINARY_DIR}/manifest.json DESTINATION ${CMAKE_INSTALL_PREFIX})
-  install(FILES webbrowser-ng.apparmor DESTINATION ${CMAKE_INSTALL_PREFIX})
+  install(FILES morph-browser.apparmor DESTINATION ${CMAKE_INSTALL_PREFIX})
 endif()
