+++ src/app/webbrowser/morph-browser.apparmor
