+++ src/app/webbrowser/morph-browser.desktop.in.in
@@ -5,23 +5,23 @@ _GenericName=Web Browser Next
 _Comment=Browse the World Wide Web using QtWebEngine
 _Keywords=Internet;WWW;Browser;Web;Explorer
 Type=Application
-Icon=@CMAKE_INSTALL_FULL_DATADIR@webbrowser-ng/webbrowser-ng.png
-Exec=webbrowser-ng %u
+Icon=@CMAKE_INSTALL_FULL_DATADIR@morph-browser/morph-browser.png
+Exec=morph-browser %u
 Terminal=false
 Categories=Network;WebBrowser;
 MimeType=text/html;text/xml;application/xhtml+xml;x-scheme-handler/http;x-scheme-handler/https;
 X-Ubuntu-Touch=true
-X-Ubuntu-Gettext-Domain=webbrowser-ng
+X-Ubuntu-Gettext-Domain=morph-browser
 X-Ubuntu-Single-Instance=true
 X-Ubuntu-Default-Department-ID=web-browsers
-X-Screenshot=@CMAKE_INSTALL_FULL_DATADIR@webbrowser-ng/screenshot.png
+X-Screenshot=@CMAKE_INSTALL_FULL_DATADIR@morph-browser/screenshot.png
 X-Ubuntu-Splash-Color=#FFFFFF
 Actions=new-window;Incognito;
 
 [Desktop Action new-window]
 _Name=New Window
-Exec=webbrowser-ng --new-window
+Exec=morph-browser --new-window
 
 [Desktop Action Incognito]
 _Name=New Private Window
-Exec=webbrowser-ng --incognito
+Exec=morph-browser --incognito
